pyproject.toml
[tool.poetry]
name = "server-task-manager"
version = "0.1.0"
description = ""
authors = ["ReykCS <reyk-carstens@web.de>"]

[tool.poetry.dependencies]
python = "^3.8"
python-socketio = {extras = ["async_client"], version = "^5.7.2"}
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
aiohttp = "^3.8.4"
bson = "^0.5.10"
pymongo = "^4.3.3"
boto3 = "^1.26.82"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
